Vagrant bootstrapping improvement
 - remove older version of openSMILE (2.1.0 -> 2.3.0)
 - add useful warnings/breakpoints inside provisioning

+++ conf/bootstrap.sh
@@ -16,7 +16,6 @@ sudo apt-get install -y git make automake libtool autoconf patch subversion fuse
     libc6-dev-i386 festival espeak python-setuptools gawk \
     libboost-all-dev
 
-
 # Kaldi and others want bash - otherwise the build process fails
 [ $(readlink /bin/sh) == "dash" ] && ln -s -f bash /bin/sh
 
@@ -24,13 +23,20 @@ sudo apt-get install -y git make automake libtool autoconf patch subversion fuse
 echo "Downloading Anaconda-2.3.0..."
 cd /home/${user}
 wget -q https://3230d63b5fc54e62148e-c95ac804525aac4b6dba79b00b39d1d3.ssl.cf1.rackcdn.com/Anaconda-2.3.0-Linux-x86_64.sh
-#bash Anaconda-2.3.0-Linux-x86_64.sh -b # batch install into /home/vagrant/anaconda
 echo "Installing Anaconda-2.3.0..."
 sudo -S -u vagrant -i /bin/bash -l -c "bash /home/${user}/Anaconda-2.3.0-Linux-x86_64.sh -b"
+
+# check if anaconda is installed correctly
+if ! [ -x "$(command -v /home/${user}/anaconda/bin/conda)" ]; then
+    echo "*******************************"
+    echo "  conda installation failed"
+    echo "*******************************"
+    exit 1
+fi
+
 if ! grep -q -i anaconda .bashrc; then
     echo "export PATH=/home/${user}/launcher:/home/${user}/utils:/home/${user}/anaconda/bin:\$PATH" >> /home/${user}/.bashrc 
 fi
-# assume 'conda' is installed now (get path)
 su ${user} -c "/home/${user}/anaconda/bin/conda install numpy scipy mkl dill tabulate joblib"
 # clean up big installer in home folder
 rm -f Anaconda-2.3.0-Linux-x86_64.sh
@@ -42,13 +48,24 @@ echo "Create python3 env"
 cd /home/$user
 cp /vagrant/conf/environment.yml /home/${user}/
 su ${user} -c "/home/${user}/anaconda/bin/conda env create -f environment.yml"
-
+if [ $? -ne 0 ]; then PYTHON3_INSTALLED=false; fi
 
 # install Matlab runtime environment
+echo "Download matlab installer"
 cd /tmp
 wget -q http://ssd.mathworks.com/supportfiles/downloads/R2017b/deployment_files/R2017b/installers/glnxa64/MCR_R2017b_glnxa64_installer.zip
+echo "Install matlab"
 unzip -q MCR_R2017b_glnxa64_installer.zip
 ./install -mode silent -agreeToLicense yes
+
+# check if matlab is installed correctly
+if [ $? -ne 0 ]; then 
+    echo "*******************************"
+    echo "  matlab installation failed"
+    echo "*******************************"
+    exit 1
+fi
+
 # add Matlab stuff to path
 echo 'LD_LIBRARY_PATH="/usr/local/MATLAB/MATLAB_Runtime/v93/runtime/glnxa64:/usr/local/MATLAB/MATLAB_Runtime/v93/bin/glnxa64:/usr/local/MATLAB/MATLAB_Runtime/v93/sys/os/glnxa64:$LD_LIBRARY_PATH"' >> /home/${user}/.bashrc
 rm /tmp/MCR_R2017b_glnxa64_installer.zip
@@ -57,19 +74,11 @@ rm /tmp/MCR_R2017b_glnxa64_installer.zip
 echo "Installing OpenSMILE"
 su ${user} -c "mkdir -p /home/${user}/repos/"
 cd /home/${user}/repos/
-wget -q http://audeering.com/download/1131/ -O OpenSMILE-2.1.tar.gz
-tar zxvf OpenSMILE-2.1.tar.gz
-# install SMILExtract system-wide
-cp openSMILE-2.1.0/bin/linux_x64_standalone_static/SMILExtract /usr/local/bin
-chmod +x /usr/local/bin/SMILExtract
-rm OpenSMILE-2.1.tar.gz
-
-# Install openSMILE2.3.0
-su ${user} -c "mkdir -p /home/${user}/repos/"
-cd /home/${user}/repos/
 wget -q https://www.audeering.com/download/1318 -O OpenSMILE-2.3.tar.gz 
 tar zxvf OpenSMILE-2.3.tar.gz
 chmod +x opensmile-2.3.0/bin/linux_x64_standalone_static/SMILExtract
+cp opensmile-2.3.0/bin/linux_x64_standalone_static/SMILExtract /usr/local/bin
+if [ $? -ne 0 ]; then OPENSMILE_INSTALLED=false; fi
 rm OpenSMILE-2.3.tar.gz
 
 # optionally Install HTK (without it, some other tools will not work)
@@ -85,10 +94,9 @@ if [ -f /vagrant/HTK.tar.gz ]; then
     sed -i "s/        /\t/g" HLMTools/Makefile # fix bad Makefile
     make all
     make install
+    if [ $? -eq 0 ]; then HTK_INSTALLED=true; fi
     fi
 fi
-
-
 
 # POPOULATE THE REPOSITORY SECTION
 cd /home/${user}/repos/
@@ -99,14 +107,11 @@ cd /home/${user}/repos/
 su ${user} -c "/home/${user}/anaconda/bin/pip install -v ipdb"
 
 cp /vagrant/conf/.theanorc /home/${user}/
-export PATH=/home/${user}/anaconda/bin:$PATH
 su ${user} -c "/home/${user}/anaconda/bin/conda install -y theano=0.8.2"
-
 
 # Install ldc-sad
 # run this version 'by hand' in the VM in repos/ using your github username and password
 #git clone http://github.com/aclew/ldc_sad_hmm
-
 
 # Install Yunitator and dependencies
 git clone https://github.com/srvk/Yunitator 
@@ -121,7 +126,6 @@ git clone https://github.com/srvk/To-Combo-SAD
 
 # Install DiarTK
 git clone http://github.com/srvk/ib_diarization_toolkit
-
 
 # Install eval
 git clone http://github.com/srvk/dscore 
@@ -138,7 +142,6 @@ python setup.py install
 #    chmod +x launcher/*
 #    git clone https://github.com/aclew/utils.git
 #    chmod +x utils/*
-
 
 # install pympi (for eaf -> rttm conversion) and tgt (for textgrid -> rttm conversion)
 # and intervaltree (needed for rttm2scp.py)
@@ -157,3 +160,20 @@ touch /home/${user}/.Xauthority
 
 # Provisioning runs as root; we want files to belong to '${user}'
 chown -R ${user}:${user} /home/${user}
+
+# Installation status 
+if ! $PYTHON3_INSTALLED; then
+    echo "*********************************************"
+    echo "Warning: python3 environment is not installed"
+    echo "*********************************************"
+fi
+if ! $OPENSMILE_INSTALLED; then
+    echo "***********************************"
+    echo "Warning: OpenSMILE is not installed"
+    echo "***********************************"
+fi
+if ! $HTK_INSTALLED; then 
+    echo "*****************************"
+    echo "Warning: HTK is not installed"
+    echo "*****************************"
+fi
